rename the module from go.etcd.io/etcd/raft to go.etcd.io/raft
