+++ raftpb/raft.proto
@@ -39,7 +39,7 @@ message Snapshot {
 }
 
 // For description of different message types, see:
-// https://pkg.go.dev/go.etcd.io/etcd/raft/v3#hdr-MessageType
+// https://pkg.go.dev/go.etcd.io/raft/v3#hdr-MessageType
 enum MessageType {
 	MsgHup             = 0;
 	MsgBeat            = 1;
