+++ rafttest/interaction_env_handler_raftstate.go
@@ -17,7 +17,7 @@ package rafttest
 import (
 	"fmt"
 
-	"go.etcd.io/etcd/raft/v3"
+	"go.etcd.io/raft/v3"
 )
 
 // isVoter checks whether node id is in the voter list within st.
