+++ rafttest/node_bench_test.go
@@ -19,7 +19,7 @@ import (
 	"testing"
 	"time"
 
-	"go.etcd.io/etcd/raft/v3"
+	"go.etcd.io/raft/v3"
 )
 
 func BenchmarkProposal3Nodes(b *testing.B) {
